requirements.txt
faster-whisper==1.0.0
ctranslate2>=4.4.0
ffmpeg-python==0.2.0
imageio-ffmpeg>=0.4.9
streamlit>=1.36.0
python-dotenv>=1.0.1
numpy>=1.26
